Removed python bindings.
The reason is that they were too heavy in size and building time, they also were very bad and not modular enough as I wanted. I'll probably add the possibility to manage object though c++ for now, and then I'll think about bindings, probably in a modular way, so people can decide to use whatever scripting language they want (or none in case of c++) to better make use of space and performance.